Move layers and file names into their own module
